Do texture copying in Start() not Awake()
There seems to be a race condition where texture loading happens in
parallel with Awake(), so copying a texture in Awake() can write a bunch
of black pixels, or just part of the texture, depending on various
conditions.  (For example screen resolution and graphics quality
settings seem to change whether the race condition is triggered.)

To fix this we move the actual copying operation to Start(), which seems
to run after all textures are loaded.