Allow user to move the H^2 view by isometries
Keys i,j,k,l move up,left,down,right in H^2.
Key r resets the view of H^2 to its original position.
Key to toggle klein/poincare changed to p.
Help updated to document p, but no documentation of i,j,k,l,r yet.

No attempt is made to Gram-Schmidt orthogonalize large isometries, nor
to return the isometry to a fundamental domain in SO(2,1), so if
a large movement is made by the user, the view of H^2 will be distorted.